// src/shared/enums/project.roles.enum.ts
export enum ProjectRole {
  // ادمین کل
  Super_Admin = "ادمین کل",
  //
  //
  //
  // مدیر فنی
  Technical_Manager = "مدیر فنی",
  //
  //
  //
  // مدیر خرید خارجی
  Foreign_Buyer_Manager = "مدیر خرید خارجی",
  // کارشناس خرید خارجی
  Foreign_Buyer_Expert = "کارشناس خرید خارجی",
  //
  //
  //
  // مدیر فروش خارجی
  Foreign_Seller_Manager = "مدیر فروش خارجی",
  // کارشناس فروش خارجی
  Foreign_Seller_Expert = "کارشناس فروش خارجی",
  //
  //
  //
  // مدیر خرید آی-سی-تی
  ICT_Buyer_Manager = "مدیر خرید آی-سی-تی",
  // کارشناس خرید آی-سی-تی
  ICT_Buyer_Expert = "کارشناس خرید آی-سی-تی",
  //
  //
  //
  // مدیر فروش آی-سی-تی
  ICT_Seller_Manager = "مدیر فروش آی-سی-تی",
  // کارشناس فروش آی-سی-تی
  ICT_Seller_Expert = "کارشناس فروش آی-سی-تی",
  //
  //
  //
  // مدیر انبار داری
  Warehouse_Manager = "مدیر انبار داری",
}